Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.1
PyJWT==2.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg[binary]==3.2.13
alembic==1.12.1
email-validator==2.1.0
httpx==0.27.0
slowapi==0.1.9
